+++ include/votca/xtp/aobasis.h
@@ -25,6 +25,7 @@
 #include "aoshell.h"
 #include "checkpoint.h"
 #include "eigen.h"
+#include <libint2/shell.h>
 
 namespace votca {
 namespace xtp {
