+++ src/tests/test_aomatrix3d.cc
@@ -24,7 +24,7 @@
 // Local VOTCA includes
 #include "votca/xtp/aomatrix.h"
 #include "votca/xtp/orbitals.h"
-
+#include <libint2/initialize.h>
 using namespace votca::xtp;
 using namespace votca;
 using namespace std;
