+++ src/tests/test_bse.cc
@@ -27,8 +27,8 @@
 #include "votca/xtp/bse.h"
 #include "votca/xtp/convergenceacc.h"
 #include "votca/xtp/qmfragment.h"
+#include <libint2/initialize.h>
 #include <votca/tools/eigenio_matrixmarket.h>
-
 using namespace votca::xtp;
 using namespace std;
