+++ src/tests/test_convergenceacc.cc
@@ -28,7 +28,7 @@
 #include "votca/xtp/aopotential.h"
 #include "votca/xtp/convergenceacc.h"
 #include "votca/xtp/orbitals.h"
-
+#include <libint2/initialize.h>
 using namespace votca::xtp;
 using namespace votca;
